[#18] refactor: 도메인과 데이터 레이어의 `photo` 모듈 이름을 `album`으로 변경
- 모듈 이름 변경에 따라 관련된 패키지 이름과 네임스페이스도 `com.chac.domain.photo`에서 `com.chac.domain.album`으로, `com.chac.data.photo`에서 `com.chac.data.album`으로 수정했습니다.
- `settings.gradle.kts`와 각 모듈의 `build.gradle.kts` 파일에서 모듈 의존성을 업데이트했습니다.
