+++ data/album/.gitignore
