+++ data/album/proguard-rules.pro
