+++ data/album/src/main/AndroidManifest.xml
