+++ data/album/src/main/java/com/chac/data/album/media/LocationUtils.kt
@@ -1,0 +1,28 @@
+package com.chac.data.album.media
+
+import android.content.Context
+import android.media.ExifInterface
+import androidx.core.net.toUri
+import com.chac.domain.album.media.MediaLocation
+import java.io.InputStream
+
+fun getMediaLocation(
+    context: Context,
+    uri: String,
+): MediaLocation? =
+    runCatching {
+        context.contentResolver.openInputStream(uri.toUri())?.use { stream ->
+            extractLocationFromStream(stream)
+        }
+    }.getOrNull()
+
+private fun extractLocationFromStream(stream: InputStream): MediaLocation? =
+    runCatching {
+        val exif = ExifInterface(stream)
+        val latLong = FloatArray(2)
+        if (exif.getLatLong(latLong)) {
+            MediaLocation(latLong[0].toDouble(), latLong[1].toDouble())
+        } else {
+            null
+        }
+    }.getOrNull()
