+++ domain/album/src/main/java/com/chac/domain/album/media/Media.kt
@@ -1,0 +1,8 @@
+package com.chac.domain.album.media
+
+data class Media(
+    val id: Long,
+    val uriString: String,
+    val dateTaken: Long,
+    val mediaType: MediaType = MediaType.IMAGE,
+)
