+++ domain/album/src/main/java/com/chac/domain/album/media/MediaLocation.kt
@@ -1,0 +1,6 @@
+package com.chac.domain.album.media
+
+data class MediaLocation(
+    val latitude: Double?,
+    val longitude: Double?,
+)
