+++ domain/album/src/main/java/com/chac/domain/album/media/MediaSortOrder.kt
@@ -1,0 +1,6 @@
+package com.chac.domain.album.media
+
+enum class MediaSortOrder {
+    NEWEST_FIRST, // 최신순
+    OLDEST_FIRST, // 오래된순
+}
