+++ domain/album/src/main/java/com/chac/domain/album/media/MediaType.kt
@@ -1,0 +1,7 @@
+package com.chac.domain.album.media
+
+enum class MediaType {
+    IMAGE,
+    VIDEO,
+    ALL,
+}
