--- domain/photo/src/main/java/com/chac/domain/photo/media/Media.kt
@@ -1,8 +1,0 @@
-package com.chac.domain.photo.media
-
-data class Media(
-    val id: Long,
-    val uriString: String,
-    val dateTaken: Long,
-    val mediaType: MediaType = MediaType.IMAGE,
-)
