--- domain/photo/src/main/java/com/chac/domain/photo/media/MediaLocation.kt
@@ -1,6 +1,0 @@
-package com.chac.domain.photo.media
-
-data class MediaLocation(
-    val latitude: Double?,
-    val longitude: Double?,
-)
