--- domain/photo/src/main/java/com/chac/domain/photo/media/MediaSortOrder.kt
@@ -1,6 +1,0 @@
-package com.chac.domain.photo.media
-
-enum class MediaSortOrder {
-    NEWEST_FIRST, // 최신순
-    OLDEST_FIRST, // 오래된순
-}
