--- domain/photo/src/main/java/com/chac/domain/photo/media/MediaType.kt
@@ -1,7 +1,0 @@
-package com.chac.domain.photo.media
-
-enum class MediaType {
-    IMAGE,
-    VIDEO,
-    ALL,
-}
