+++ feature/album/build.gradle.kts
@@ -12,7 +12,7 @@ android {
 dependencies {
     implementation(projects.core.designsystem)
     implementation(projects.core.permission)
-    implementation(projects.domain.photo)
+    implementation(projects.domain.album)
 
     implementation(libs.androidx.navigation3.runtime)
     implementation(libs.androidx.lifecycle.viewmodel.ktx)
