+++ feature/album/src/main/java/com/chac/feature/album/clustering/ClusteringViewModel.kt
@@ -4,8 +4,8 @@ import androidx.compose.runtime.Immutable
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.chac.domain.photo.media.Media
-import com.chac.domain.photo.media.MediaRepository
+import com.chac.domain.album.media.Media
+import com.chac.domain.album.media.MediaRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
